refactor: rename JS defaults to camelCase filenames
accessibility-checker's JS source uses camelCase filenames throughout
(preventLinksOpeningNewWindowFix.js, useIsPostEditor.js, etc.); EDMM's
src/js/defaults/ was the one holdout using kebab-case. Rename
render-info.js -> renderInfo.js and render-pagination.js ->
renderPagination.js, update the two import sites in instance.js.
