+++ src/js/defaults/renderInfo.js
