+++ src/js/defaults/renderPagination.js
